backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
sqlparse==0.5.3
typing_extensions==4.14.1
psycopg2-binary==2.9.9
whitenoise==6.6.0
gunicorn==21.2.0
dj-database-url==2.1.0
Pillow==10.4.0
setuptools==69.5.1
pillow-heif==0.18.0
python-dotenv==1.0.0

# Testing dependencies
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1

# Security dependencies for file validation
python-magic==0.4.27
pillow-simd==10.0.1
